image.py
from openai import OpenAI

from dotenv import load_dotenv

load_dotenv(".env",override=True)

client = OpenAI()


response = client.images.generate(
    prompt="A cute baby sea otter",
    n=2,
    size="1024x1024"
)

print(response.data[0].url)