requirements.txt
openai
python-dotenv
langchain_community
langchain_openai
huggingface_hub